Revert "Liquid staking page"
